+++ src/examples/lang_cpp/operator_new.cc
@@ -46,10 +46,10 @@
  * being called.
  */
 
-inline void *operator new(const unsigned int size) {
+inline void *operator new(const size_t size) {
 	void *p=malloc(size);
 
-	fprintf(stderr, "in operator new with size=%d,p=%p\n", size, p);
+	fprintf(stderr, "in operator new with size=%zd,p=%p\n", size, p);
 	return(p);
 }
 
@@ -58,17 +58,17 @@ inline void operator delete(void *p) {
 	free(p);
 }
 
-inline void *operator new(const unsigned int size, const unsigned int type) {
+inline void *operator new(const size_t size, const unsigned int type) {
 	void *p=malloc(size);
 
-	fprintf(stderr, "in operator new with size=%d,type=%d,p=%p\n", size, type, p);
+	fprintf(stderr, "in operator new with size=%zd,type=%d,p=%p\n", size, type, p);
 	return(p);
 }
 
-inline void *operator new[] (const unsigned int size) {
+inline void *operator new[] (const size_t size) {
 	void *p=malloc(size);
 
-	fprintf(stderr, "in operator new[] with size=%d, p=%p\n", size, p);
+	fprintf(stderr, "in operator new[] with size=%zd, p=%p\n", size, p);
 	return(p);
 }
 
@@ -77,10 +77,10 @@ inline void operator delete[] (void *p) {
 	free(p);
 }
 
-inline void *operator new[] (const unsigned int size, const unsigned int type) {
+inline void *operator new[] (const size_t size, const unsigned int type) {
 	void *p=malloc(size);
 
-	fprintf(stderr, "in operator new[] with size=%d,type=%d,p=%p\n", size, type, p);
+	fprintf(stderr, "in operator new[] with size=%zd,type=%d,p=%p\n", size, type, p);
 	return(p);
 }
